pyproject.toml
[project]
name = "clari-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.20",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "tiktoken>=0.11.0",
    "typer>=0.16.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
